done: A a p f. in progress: F R G C 1 x r
