+++ mx_process_arg.c
@@ -28,7 +28,8 @@ void mx_process_arg(char **args, t_flags *flags) {
     *entries = NULL;
     while (temp_args[i])
         mx_save_info(temp_args[i++], flags, stats, entries);
-    mx_sort_lists(flags, stats, entries);                   //TODO
+    if (!flags->f)
+        mx_sort_lists(flags, stats, entries);
     mx_output(flags, stats, entries, args);
     free_lists(stats, entries);
 }
