+++ src/mx_output.c
@@ -4,10 +4,10 @@ void mx_output(t_flags *flags,
                t_list **stats, 
                t_list **entries, 
                char **args) {
-    //t_list *temp_stat = *stats;
     (void)stats;
     (void)flags;
     t_list *temp_entry = *entries;
+    t_list *temp_stat = *stats;
     static bool first = true;
 
     if (!first) {
@@ -20,8 +20,32 @@ void mx_output(t_flags *flags,
     first = false;
     while(temp_entry) {
         struct dirent *temp_dirent = temp_entry->data;
-        mx_printstr(temp_dirent->d_name);
-        mx_printchar('\n');
+        struct stat *temp_st = temp_stat->data;
+
+        if (flags->A) {
+            if (!mx_strcmp(temp_dirent->d_name, ".") 
+             || !mx_strcmp(temp_dirent->d_name, "..")) {
+                temp_entry = temp_entry->next;
+                continue;
+             }
+        }
+        if (mx_strlen(temp_dirent->d_name) > 0) {   //TODO: unprinted characters
+            mx_printstr(temp_dirent->d_name);
+            if (flags->F || flags->p) {
+            if (temp_dirent->d_type == DT_DIR)
+                mx_printchar('/');
+            else if (flags->F && temp_dirent->d_type == DT_LNK)
+                mx_printchar('@');
+            else if (flags->F && temp_dirent->d_type == DT_SOCK)
+                mx_printchar('=');
+            else if (flags->F && temp_dirent->d_type == DT_FIFO)
+                mx_printchar('|');
+            else if (flags->F && temp_st && temp_st->st_mode & S_IXUSR)
+                mx_printchar('*');
+            }
+            mx_printchar('\n');
+        }
         temp_entry = temp_entry->next;
+        temp_stat = temp_stat->next;
     }
 }
